fix(sim): store the action scale at full precision in the dataset
Saving it as float32 turned 0.025 into 0.02500000037252903, which then
travelled through the checkpoint into the ONNX metadata string and the
manifest's action_scale. The value is provenance, not training data, so it is
stored as float64.

+++ sim/collect_demos.py
@@ -84,7 +84,9 @@ def main() -> None:
         actions=actions,
         episode_ends=episode_ends,
         success=successes,
-        action_scale=np.float32(args.action_scale),
+        # float64: a float32 round-trip turns 0.025 into 0.02500000037252903,
+        # which then reaches the ONNX metadata and the manifest as that string.
+        action_scale=np.float64(args.action_scale),
     )
 
     print(
